Add audit chain enforcement and tests
